proga/requirements.txt.txt
PyQt5==5.15.10
psycopg2-binary==2.9.9